restructured test to utils
